refactor: move domain types and sort utils out of shared/
- Move FolderEntry to folder-navigation/types/folderTypes.ts
- Move createFolderEntry to folder-navigation/utils/folderEntry.ts
- Move FolderSortFunction and naturalFolderSort to folder-navigation
- Move SortFunction and naturalSort to image-viewer
- Remove inverted shared/types dependencies on feature modules
- Colocate sort utility tests with their owning features
- Trim shared/index.ts to feature-agnostic exports only
