+++ src/features/image-viewer/types/SortFunction.ts
@@ -1,0 +1,6 @@
+import type { ImageSource } from './ImageSource';
+
+/**
+ * ImageSource 配列をソートするための関数型。
+ */
+export type SortFunction = (a: ImageSource, b: ImageSource) => number;
