+++ src/features/image-viewer/utils/sort.ts
@@ -3,7 +3,5 @@ import type { SortFunction } from '../types/SortFunction';
 export const naturalSort: SortFunction = (a, b) => {
   const [aName, bName] = [a.name, b.name];
 
-  const result = aName.localeCompare(bName, 'ja', { numeric: true });
-
-  return result;
+  return aName.localeCompare(bName, 'ja', { numeric: true });
 };
