--- src/shared/types/FolderSortFunction.ts
@@ -1,19 +1,0 @@
-import type { FolderEntry } from '../../features/folder-navigation/hooks/useSiblingContainers';
-
-/**
- * FolderSortFunction
- * @description
- * FolderSortFunction is a type that represents a function used to sort an array of FolderEntry objects.
- * It takes two FolderEntry objects as arguments and returns a number indicating their relative order.
- * A negative number indicates that the first argument should come before the second,
- * a positive number indicates that the first argument should come after the second,
- * and zero indicates that they are equal in terms of sorting order.
- *
- * (JP)
- * FolderSortFunctionは、FolderEntryオブジェクトの配列をソートするために使用される関数を表す型です。
- * 2つのFolderEntryオブジェクトを引数として受け取り、それらの相対的な順序を示す数値を返します。
- * 1. 負の数は、最初の引数が2番目の引数の前に来るべきことを示します。
- * 2. 正の数は、最初の引数が2番目の引数の後に来るべきことを示します。
- * 3. ゼロは、ソート順序において等しいことを示します。
- */
-export type FolderSortFunction = (a: FolderEntry, b: FolderEntry) => number;
